+++ public/sw.js
@@ -1,4 +1,4 @@
-const OED_VERSION = '2.0.20200417.0';
+const OED_VERSION = '2.0.20200417.1';
 const OED_BASE = 'OED';
 const ACE_VERSION = '1.4.10';
 const ACE_BASE = 'Ace';
@@ -29,7 +29,7 @@ const OED_URLS = [
     './dialog_fold_style.js',
     './dialog_font_family.js',
     './dialog_font_size.js',
-    './dialog_keybinding.js',
+    './dialog_keyboard_handler.js',
     './dialog_language_mode.js',
     './dialog_merge_undo_deltas.js',
     './dialog_new_line_mode.js',